requirements.txt
djangorestframework >= 3.7.3
django >= 1.11
Pillow >= 4.3.0